Headers updated, image now shows as headerStyle conflicts with headerbackground fix | Also, moved colors.ts to utils, many files changed due to imports
